pyproject.toml
[tool.poetry]
name = "rust"
version = "0.1.0"
description = "Rust support for Biscuit"
authors = ["Billy <billydevbusiness@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
biscuit-editor = "^2.99.99"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
